backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
bcrypt==4.0.1
passlib==1.7.4
cryptography==41.0.7
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
email-validator==2.0.0
alembic==1.12.1
python-dotenv==1.0.0
pydantic-settings==2.1.0
prometheus-client
python-json-logger
psycopg2-binary==2.9.9